Improve services utilization (#7)
# Goal
The goal was to improve the services utilization and follow the new way of using this kind of library. We are now completely able to inject our services in an IoC, resolve them and use them more easily than before. Because of that, we completely broke the compatibility.

We also took this opportunity to use IReadOnlyCollection instead of IEnumerable. This was made to simplify the user library so that he should not always enumerate the list after a request.

# Technical aspects
All the provider class are internal and can only be created by injecting them in an IoC. The models, providers and extensions are still available. To inject those services, use the ServiceCollectionExtensions.AddFootballDataService(...) method.

Also simplified the providers so that we don't have the base address everywhere.