Adapt models to API V4 (#8)
# Goal

The goal was to start the rework of the library so that we will be able to use the [API V4](https://www.football-data.org/). The first step was to rework completely the objects that will be transmitted to the end user.

# Technical aspects

A lot of similar/duplicated class were created to be as much accurate as we can with the received information of the API. Even if some objects are very similar I preferred create an object for each representation (some of them I've made them inherit from the others).

Currently the unit tests have been removed because totally outdated and should completely be reworked to match the new standards.