Rework providers to be more restrictive (#9)
# Goal

The goal of this PR is to propose a new way to get the information from the Web API. Before the methods were not that clear on what you can do and how you can filter the values.
This new implementation will help the user to use the right methods with the appropriate filters. Documentation of the providers must follow in future PRs

# Technical aspects

Based on the [API documentation](https://docs.football-data.org/general/v4/resources.html), I started to reimplement and to reorganize the providers that were already existing.
We started by renaming the methods to follow Microsoft recommendations (`Async` at the end of asynchronous method), but the biggest change comes with the arguments rework.

Now I analyzed a bit more the documentation and created multiple methods were you can pass optional filters. This will probably help more the user with filters when he will request data. Nevertheless, documentation of the API must still follow.